01 done

+++ Year 3/02. Recursion/01. Fibonacci/Program.cs
@@ -4,12 +4,12 @@
     {
         static void Main(string[] args)
         {
-            int input = int.Parse(Console.ReadLine());
+            ulong input = ulong.Parse(Console.ReadLine());
 
             Console.WriteLine(Fiboncci(input));
         }
 
-        public static int Fiboncci(int n)
+        public static ulong Fiboncci(ulong n)
         {
             if (n <= 2)
             {
